blog_site/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.1.5
django-import-export==2.0.2
django-registration-redux==2.7
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.2
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
pytz==2020.5
PyYAML==5.3
requests==2.25.1
somewhere==0.0.1
sqlparse==0.4.1
tablib==1.1.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
